Move code from Navigation.cs into ActionNavigation class
